requirements.txt
efficientnet-pytorch==0.7.1
matplotlib==3.5.2
numpy==1.21.5
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.3.5
Pillow==9.2.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
protobuf==3.20.1
